backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
langchain>=0.1.0
langchain-ollama>=0.1.0
ddgs>=1.0.0
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4
playwright>=1.40.0
python-multipart>=0.0.9
aiofiles>=23.0.0
python-dotenv>=1.0.1
sarvamai>=0.1.0
